지정된 시간에 알람울리기 1~50까지의 숫자중에서 버튼을 눌러야 하는 횟수 지정

+++ KoKoAlarm/src/hs/mirim/CustomArrayAdapter.java
@@ -6,6 +6,7 @@ import android.app.Activity;
 import android.app.AlarmManager;
 import android.app.PendingIntent;
 import android.content.Context;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.View.OnClickListener;
@@ -13,7 +14,8 @@ import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
 
-public class CustomArrayAdapter extends ArrayAdapter<Alarm>{	
+public class CustomArrayAdapter extends ArrayAdapter<Alarm>{
+	final static String TAG = "꼬꼬:CustomArryayAdapter";
 	
 	Alarm alarm;
 	public CustomArrayAdapter(Context listSampleActivity,
@@ -30,14 +32,19 @@ public class CustomArrayAdapter extends ArrayAdapter<Alarm>{
 		}
 	
 		alarm=getItem(position);
+		Log.e(TAG, "알람 id는" + alarm.getId() );
 	
 		Button btnDelete = (Button)view.findViewById(R.id.btnDelete);
 		btnDelete.setOnClickListener(new OnClickListener() {
-	
-			
 		public void onClick(View arg0) {
+			if(alarm != null){
+			Log.e(TAG, "지우려고 하는 알람 id는" + alarm.getId() );
+				
 			alarm.delete();
 			((AddAlarm) getContext()).deleteItem(alarm);
+			}else{
+				Log.e(TAG, "지울 알람 객체가 없습니다.");
+			}
 		}
 	});
